--- a/.gitmodules
+++ b/.gitmodules
@@ -1,7 +1,3 @@
-[submodule "deps/OpenBLAS"]
-	path = deps/OpenBLAS
-	url = https://github.com/xianyi/OpenBLAS.git
-
 [submodule "deps/pybind11"]
 	path = deps/pybind11
 	url = https://github.com/pybind/pybind11.git