--- a/src/VecSim/algorithms/brute_force.cpp
+++ b/src/VecSim/algorithms/brute_force.cpp
@@ -8,7 +8,7 @@
 #include <limits>
 #include "VecSim/spaces/space_interface.h"
 #include "VecSim/spaces/L2_space.h"
-#include "VecSim/spaces/internal_product.h"
+#include "VecSim/spaces/IP_space.h"
 #include "VecSim/utils/arr_cpp.h"
 #include <iostream>