--- a/src/VecSim/algorithms/hnsw/hnswlib.h
+++ b/src/VecSim/algorithms/hnsw/hnswlib.h
@@ -12,7 +12,7 @@
 #include <sys/resource.h>
 
 #include "visited_list_pool.h"
-#include "VecSim/spaces/L2.h"
+#include "VecSim/spaces/L2_space.h"
 #include "VecSim/spaces/internal_product.h"
 #include "VecSim//spaces/space_interface.h"
 #include "VecSim/utils/arr_cpp.h"