--- a/src/VecSim/algorithms/hnsw/hnswlib_c.cpp
+++ b/src/VecSim/algorithms/hnsw/hnswlib_c.cpp
@@ -4,7 +4,7 @@
 #include <cassert>
 #include "VecSim/utils/arr_cpp.h"
 #include "VecSim/algorithms/hnsw/hnswlib.h"
-#include "VecSim/spaces/L2.h"
+#include "VecSim/spaces/L2_space.h"
 #include "VecSim/spaces/internal_product.h"
 
 using namespace std;