--- a/src/VecSim/algorithms/hnsw/hnswlib_c.cpp
+++ b/src/VecSim/algorithms/hnsw/hnswlib_c.cpp
@@ -5,7 +5,7 @@
 #include "VecSim/utils/arr_cpp.h"
 #include "VecSim/algorithms/hnsw/hnswlib.h"
 #include "VecSim/spaces/L2_space.h"
-#include "VecSim/spaces/internal_product.h"
+#include "VecSim/spaces/IP_space.h"
 
 using namespace std;
 using namespace hnswlib;