--- a/src/VecSim/spaces/IP/IP.h
+++ b/src/VecSim/spaces/IP/IP.h
@@ -1,0 +1,3 @@
+#pragma once
+
+float InnerProduct(const void *pVect1, const void *pVect2, const void *qty_ptr);