--- a/src/VecSim/spaces/IP/IP_AVX512.h
+++ b/src/VecSim/spaces/IP/IP_AVX512.h
@@ -1,0 +1,5 @@
+#pragma once
+
+float InnerProductSIMD16Ext_AVX512(const void *pVect1v, const void *pVect2v, const void *qty_ptr);
+float InnerProductSIMD16ExtResiduals_AVX512(const void *pVect1v, const void *pVect2v,
+                                            const void *qty_ptr);