--- a/src/VecSim/spaces/IP/IP_SSE.h
+++ b/src/VecSim/spaces/IP/IP_SSE.h
@@ -1,0 +1,8 @@
+#pragma once
+
+float InnerProductSIMD4Ext_SSE(const void *pVect1v, const void *pVect2v, const void *qty_ptr);
+float InnerProductSIMD16Ext_SSE(const void *pVect1v, const void *pVect2v, const void *qty_ptr);
+float InnerProductSIMD4ExtResiduals_SSE(const void *pVect1v, const void *pVect2v,
+                                        const void *qty_ptr);
+float InnerProductSIMD16ExtResiduals_SSE(const void *pVect1v, const void *pVect2v,
+                                         const void *qty_ptr);