--- a/src/VecSim/spaces/L2.cpp
+++ b/src/VecSim/spaces/L2.cpp
@@ -158,7 +158,6 @@ static float L2SqrSIMD4ExtResiduals(const void *pVect1v, const void *pVect2v, co
 }
 #endif
 
-using namespace hnswlib;
 
 L2Space::L2Space(size_t dim) {
     fstdistfunc_ = L2Sqr;