--- a/src/VecSim/spaces/L2.h
+++ b/src/VecSim/spaces/L2.h
@@ -2,7 +2,6 @@
 
 #include "space_interface.h"
 
-namespace hnswlib {
 class L2Space : public SpaceInterface<float> {
 
     DISTFUNC<float> fstdistfunc_;
@@ -17,4 +16,3 @@ class L2Space : public SpaceInterface<float> {
     DISTFUNC<float> get_dist_func() const override;
     void *get_data_dim() override;
 };
-} // namespace hnswlib