--- a/src/VecSim/spaces/L2/L2.h
+++ b/src/VecSim/spaces/L2/L2.h
@@ -1,0 +1,3 @@
+#pragma once
+
+float L2Sqr(const void *pVect1v, const void *pVect2v, const void *qty_ptr);