--- a/src/VecSim/spaces/L2/L2_AVX.h
+++ b/src/VecSim/spaces/L2/L2_AVX.h
@@ -1,0 +1,5 @@
+#pragma once
+
+float L2SqrSIMD16Ext_AVX(const void *pVect1v, const void *pVect2v, const void *qty_ptr);
+
+float L2SqrSIMD16ExtResiduals_AVX(const void *pVect1v, const void *pVect2v, const void *qty_ptr);