--- a/src/VecSim/spaces/L2/L2_AVX512.cpp
+++ b/src/VecSim/spaces/L2/L2_AVX512.cpp
@@ -1,0 +1,49 @@
+#include "L2_AVX512.h"
+#include "L2.h"
+#include "VecSim/spaces/space_includes.h"
+#ifdef __AVX512__
+
+float L2SqrSIMD16Ext_AVX512(const void *pVect1v, const void *pVect2v, const void *qty_ptr) {
+    float *pVect1 = (float *)pVect1v;
+    float *pVect2 = (float *)pVect2v;
+    size_t qty = *((size_t *)qty_ptr);
+    float PORTABLE_ALIGN32 TmpRes[16];
+    size_t qty16 = qty >> 4;
+
+    const float *pEnd1 = pVect1 + (qty16 << 4);
+
+    __m512 diff, v1, v2;
+    __m512 sum = _mm512_set1_ps(0);
+
+    while (pVect1 < pEnd1) {
+        v1 = _mm512_loadu_ps(pVect1);
+        pVect1 += 16;
+        v2 = _mm512_loadu_ps(pVect2);
+        pVect2 += 16;
+        diff = _mm512_sub_ps(v1, v2);
+        // sum = _mm512_fmadd_ps(diff, diff, sum);
+        sum = _mm512_add_ps(sum, _mm512_mul_ps(diff, diff));
+    }
+
+    _mm512_store_ps(TmpRes, sum);
+    float res = TmpRes[0] + TmpRes[1] + TmpRes[2] + TmpRes[3] + TmpRes[4] + TmpRes[5] + TmpRes[6] +
+                TmpRes[7] + TmpRes[8] + TmpRes[9] + TmpRes[10] + TmpRes[11] + TmpRes[12] +
+                TmpRes[13] + TmpRes[14] + TmpRes[15];
+
+    return (res);
+}
+
+float L2SqrSIMD16ExtResiduals_AVX512(const void *pVect1v, const void *pVect2v,
+                                     const void *qty_ptr) {
+    size_t qty = *((size_t *)qty_ptr);
+    size_t qty16 = qty >> 4 << 4;
+    float res = L2SqrSIMD16Ext_AVX512(pVect1v, pVect2v, &qty16);
+    float *pVect1 = (float *)pVect1v + qty16;
+    float *pVect2 = (float *)pVect2v + qty16;
+
+    size_t qty_left = qty - qty16;
+    float res_tail = L2Sqr(pVect1, pVect2, &qty_left);
+    return (res + res_tail);
+}
+
+#endif