--- a/src/VecSim/spaces/L2/L2_AVX512.h
+++ b/src/VecSim/spaces/L2/L2_AVX512.h
@@ -1,0 +1,5 @@
+#pragma once
+
+float L2SqrSIMD16Ext_AVX512(const void *pVect1v, const void *pVect2v, const void *qty_ptr);
+
+float L2SqrSIMD16ExtResiduals_AVX512(const void *pVect1v, const void *pVect2v, const void *qty_ptr);