… src/CMakeLists.txt
… src/VecSim/spaces/IP_space.h
… src/VecSim/spaces/L2/L2_SSE.cpp
… src/VecSim/spaces/L2/L2_SSE.h
… src/VecSim/spaces/L2_space.cpp
… src/VecSim/spaces/L2_space.h
… src/VecSim/spaces/internal_product.cpp
… src/VecSim/spaces/internal_product.h
… src/VecSim/spaces/space_aux.cpp
… src/VecSim/spaces/space_aux.h
… src/VecSim/spaces/space_includes.h
… src/VecSim/spaces/space_interface.h
… src/VecSim/utils/vec_utils.cpp
… src/VecSim/utils/vec_utils.h
… tests/unit/test_bruteforce.cpp
… tests/unit/test_hnswlib.cpp