This PR:
1. Removes OpenBLAS and introduces native vector operations for SSE, AVX and AVX512 for both IP and L2 spaces
2. Solves a bug where `inf` is the score in brute force algorithm
3. as a result of 1+2 brute force algorithm search for top-k has been modified and without noticeable performance lost:

Before: Benchmarked simplified version of brute force algorithm via Google Bench with 1M*128 vectors of FP32
OpenBlas_top_10_ip           43784172 ns  ~ 43 ms
After: Benchmarked viae2e test of the python bindings:
lookup time for 1M *128 vectors took 0.04492020606994629 seconds ~ 44ms